// my-frontend/src/types/order.ts
export interface MenuItem {
  name: string;
  price: number;
  img: string;
}

export interface OrderFormData {
  customer_Name: string;
  orderType: 'Dine In' | 'Take Out';
  deliveryType: 'Restaurant' | 'Room Delivery';
  roomNumber?: string;
  items: Map<string, number>;
  paymentMethod: string;
  AccountName: string;
  referenceNumber: string;
}

export interface Sales {
  id?: number; // This will be auto-generated by the database
  itemName: string; // Item name for the sale
  quantity: number; // Quantity of the item sold
  totalAmount: number; // Total amount of the sale
  saleDate: string; // The date when the sale was made (ISO string or your preferred format)
  createdAt?: string; // Automatically populated by the database
  updatedAt?: string; // Automatically populated by the database
}


export interface OrderSummary extends OrderFormData {
  total: number;
}